requirements.txt
requests
beautifulsoup4
pandas
tqdm
fake-useragent
python-dotenv
pyyaml
dnspython
selenium
undetected-chromedriver
openpyxl